+++ solution/tasks event listeners.js
@@ -1,4 +1,6 @@
-import { localStorageSave } from './localStorage';
+import {
+  localStorageSave, toDoTasksUl, inProgressTasksUl, doneTasksUl
+} from './localStorage';
 import listCounter from './list counter for index';
 
 // gaining focus function
@@ -25,4 +27,40 @@ export function endDrag (e) {
   e.target.classList.remove('dragging');
   localStorageSave();
   listCounter();
+}
+
+export function saveValueBlur (e) {
+  const { target } = e;
+  if (target.tagName !== 'LI') {
+    return;
+  }
+  target.setAttribute('contenteditable', 'false');
+  target.style.backgroundColor = 'rgba(0,0,0,0)';
+  localStorageSave();
+  listCounter();
+}
+
+function hoverReplace (e) {
+  const { target } = e;
+  function innerKeyReplace (ev) {
+    if (ev.altKey) {
+      checkListAtAlt(ev.key, target, 1, toDoTasksUl[0]);
+      checkListAtAlt(ev.key, target, 2, inProgressTasksUl[0]);
+      checkListAtAlt(ev.key, target, 3, doneTasksUl[0]);
+    }
+    // local storage insertion
+    localStorageSave();
+    listCounter();
+  }
+  target.addEventListener('mouseleave', () => {
+    window.removeEventListener('keydown', innerKeyReplace);
+  });
+  window.addEventListener('keydown', innerKeyReplace);
+}
+
+export function addHoverReplace (e) {
+  if (e.target.tagName !== 'LI') {
+    return;
+  }
+  e.target.addEventListener('mouseenter', hoverReplace);
 }
